# dashboard/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gemstar / BodySpace API",
    "version": "1.0.0",
    "description": "API for the BodySpace Recovery Studio marketing automation platform. Protected routes under `/api/bodyspace` require an active Azure AD session."
  },
...
  "components": {
    "schemas": {
      "AvailabilitySignal": {
        "type": "string",
        "enum": [
          "push",
          "hold",
          "pause"
        ],
        "description": "Booking availability signal"
      },
      "CampaignStatus": {
        "type": "string",
        "enum": [
          "draft",
          "pending_review",
          "approved",
          "rejected",
          "scheduled",
          "published"
        ],
        "description": "Current lifecycle status of a campaign"
      },
      "PostStatus": {
        "type": "string",
        "enum": [
          "draft",
          "pending_review",
          "approved",
          "rejected",
          "scheduled",
          "published"
        ],
        "description": "Current lifecycle status of a social post"
      },
      "ImageStatus": {
        "type": "string",
        "enum": [
          "needed",
          "generating",
          "draft",
          "approved"
        ],
        "description": "Image generation status"
      },
      "Platform": {
        "type": "string",
        "enum": [
          "instagram",
          "facebook"
        ],
        "description": "Social media platform"
      },
      "PostType": {
        "type": "string",
        "enum": [
          "feed",
          "story",
          "reel"
        ],
        "description": "Type of social media post"
      },
      "ContentPillar": {
        "type": "string",
        "enum": [
          "education",
          "promotion",
          "community",
          "social_proof",
          "seasonal"
        ],
        "description": "Content strategy pillar"
      },
      "SocialPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "post_abc123"
          },
          "campaignId": {
            "type": "string",
            "example": "cmp_xyz789"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "postType": {
            "$ref": "#/components/schemas/PostType"
          },
          "contentPillar": {
            "$ref": "#/components/schemas/ContentPillar"
          },
          "copy": {
            "type": "string",
            "description": "Post body text"
          },
          "ownerEdit": {
            "type": "string",
            "description": "Owner override for post copy"
          },
          "imageDirection": {
            "type": "string",
            "description": "Instructions for image generation"
          },
          "hashtags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "callToAction": {
            "type": "string"
          },
          "scheduledFor": {
            "type": "string",
            "example": "2025-06-01T09:00:00+08:00"
          },
          "status": {
            "$ref": "#/components/schemas/PostStatus"
          },
          "imageUrl": {
            "type": "string",
            "description": "Publicly accessible URL of the generated image"
          },
          "imageStatus": {
            "$ref": "#/components/schemas/ImageStatus"
          },
          "rejectionReason": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "example": "2025-05-20T12:00:00.000Z"
          },
          "publishedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "campaignId",
          "platform",
          "postType",
          "contentPillar",
          "copy",
          "imageDirection",
          "hashtags",
          "callToAction",
          "status",
          "createdAt"
        ]
      },
      "Campaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmp_xyz789"
          },
          "name": {
            "type": "string",
            "example": "Winter Wellness Push"
          },
          "theme": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "targetServices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "durationWeeks": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          },
          "freshaSignals": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "signal": {
                  "$ref": "#/components/schemas/AvailabilitySignal"
                },
                "slots": {
                  "type": "integer"
                }
              },
              "required": [
                "signal",
                "slots"
              ]
            }
          },
          "trendsBriefId": {
            "type": "string"
          },
          "ownerNotes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "approvedAt": {
            "type": "string"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SocialPost"
            }
          }
        },
        "required": [
          "id",
          "name",
          "theme",
          "description",
          "targetServices",
          "durationWeeks",
          "status",
          "freshaSignals",
          "createdAt",
          "posts"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/bodyspace/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "List campaigns, optionally filtered by status",
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/CampaignStatus"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "campaigns": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Campaign"
                      }
                    }
                  },
                  "required": [
                    "ok",
                    "campaigns"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bodyspace/campaigns/{id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get a single campaign by ID",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "campaign": {
                      "$ref": "#/components/schemas/Campaign"
                    }
                  },
                  "required": [
                    "ok",
                    "campaign"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bodyspace/wizard/campaign": {
      "post": {
        "tags": [
          "Wizard"
        ],
        "summary": "Run the campaign wizard to generate a new campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownerBrief": {
                    "type": "string"
                  },
                  "selectedServices": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "campaign": {
                      "$ref": "#/components/schemas/Campaign"
                    }
                  },
                  "required": [
                    "ok",
                    "campaign"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bodyspace/run/campaign": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Manually trigger the campaign planner agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownerBrief": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}